Fixed one warning

+++ lib.cpp
@@ -7938,13 +7938,11 @@ void ChartFrame::GetCoastLineData_Mercator(void){
         
         for(j=j_min; j<j_max; j++){
             
-            
-            
             //            cout << "\nCalled data_x[" << i - floor_min_lat << "][" << j % 360;
             //            flush(cout);
             
             //count how many datapoints are in data_x[i][j] and in data_y[i][j]
-            n = ((parent->data_x)[i - floor_min_lat][j % 360]).size();
+            n = ((unsigned int)(((parent->data_x)[i - floor_min_lat][j % 360]).size()));
             
             every = (unsigned int)(((double)n)/((double)(((parent->plot)->n_points_plot_coastline).value))*((double)n_points_grid));
             if(every == 0){every = 1;}
